requirements.txt
asn1crypto==1.5.1
loguru==0.7.2
pycryptodome==3.19.0
